Move to base64 encoded data
