+++ nostalgia@asphyxia/data/.gitignore
@@ -1,3 +1,3 @@
 *.json
-!forte_mdb.json
-forte_mdb.xml
+*.json
+*.xml
